+++ packages/open-api-client/examples/real-world-apis/github/next.functional.ts
@@ -6,7 +6,7 @@ import {
   type OpenAPIClientConfig,
   openAPIFetch,
   type OpenAPIResponses,
-} from "jsr:@brad-jones/deno-net-open-api-client@0.2.0";
+} from "jsr:@brad-jones/deno-net-open-api-client@0.2.1";
 import { z } from "npm:zod@^4.1.12";
 
 export type ActionsArtifactAndLogRetention = {
@@ -168951,7 +168951,7 @@ export const metaGetZen = (config: OpenAPIClientConfig): Promise<
     },
   ) as any;
 
-export { createCustomClient } from "jsr:@brad-jones/deno-net-open-api-client@0.2.0";
+export { createCustomClient } from "jsr:@brad-jones/deno-net-open-api-client@0.2.1";
 
 /**
  * Create a client with all available operations.