+++ packages/open-api-client/examples/real-world-apis/paypal/payments_payment_v2.functional.ts
@@ -6,7 +6,7 @@ import {
   type OpenAPIClientConfig,
   openAPIFetch,
   type OpenAPIResponses,
-} from "jsr:@brad-jones/deno-net-open-api-client@0.2.0";
+} from "jsr:@brad-jones/deno-net-open-api-client@0.2.1";
 import { z } from "npm:zod@^4.1.12";
 
 /** Captures either a portion or the full authorized amount of an authorized payment. */
@@ -1762,7 +1762,7 @@ export const refundsGet = (config: OpenAPIClientConfig, request: {
     request,
   ) as any;
 
-export { createCustomClient } from "jsr:@brad-jones/deno-net-open-api-client@0.2.0";
+export { createCustomClient } from "jsr:@brad-jones/deno-net-open-api-client@0.2.1";
 
 /**
  * Create a client with all available operations.